feat: reduce reconciliation deltas
This was done by adding a server side quantizer for client update data
events. This means that most of the time, on each server tick we are
able to process a different client time bundle, so that the server
behavior is quite similar to the client behavior.

Aditionally the client now only sends out when a bundle is ready, before
there was a periodic signal that would desync from the bundle creation
causing there to be times when no bundles to be sent. This also improves
reconciliation detlas which stemmed from the fact that a bundle wouldn't
make it out "in time".